requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
elasticsearch==8.11.0
redis==5.0.1
pydantic==2.12.3
python-multipart==0.0.6
python-dotenv==1.0.0
aiohttp==3.9.1
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
